Refactor settings import paths and remove deprecated meta files
