+++ app/robots.ts
@@ -1,6 +1,6 @@
 import type { MetadataRoute } from "next"
 
-import { Settings } from "@/lib/meta"
+import { Settings } from "@/types/settings"
 
 export default function robots(): MetadataRoute.Robots {
   return {
