+++ lib/transition/browser-native.ts
